Remove alerting.go and move to alert.go
